add authorisation

+++ index.php
@@ -98,26 +98,93 @@ get('/resources', function () {
 
 
 //get
-get('/resources/admin', function () {	render('resources_admin', array('title' => "Edit resources"));});
+get('/resources/admin', function () {	
 
-get('/resources/admin/new', function () {	render('res_new', array('title' => "Add resources"));});
+	//check session
+	$authorized = check_session();
+
+	if($authorized){
+		render('res_main', array('title' => "Manage Resources"));
+	} else {
+		render('res_login', array('title' => "Login"));
+	}
+
+});
+
+get('/resources/admin/new', function () {	
+
+	//check session
+	$authorized = check_session();
+
+	if($authorized){
+		render('res_new', array('title' => "Add resources"));
+	} else {
+		render('res_login', array('title' => "Login", 'error' => 'you must login to view this page.'));
+	}
+
+});
 
 get('/resources/admin/edit', function () {	
 
-	$files = get_files();
-	render('res_edit', array('title' => "Edit Resources:",
-							  'files' => $files ));
+	//check session
+	$authorized = check_session();
+
+	if($authorized){
+		$files = get_files();
+		render('res_edit', array('title' => "Edit Resources:",
+								  'files' => $files ));
+	} else {
+		render('res_login', array('title' => "Login", 'error' => 'you must login to view this page.'));
+	}
 });
 
 get('/resources/admin/delete', function () {	
 
-	$files = get_files();
-	render('res_delete', array('title' => "Delete Resources:",
-							  'files' => $files ));
+	//check session
+	$authorized = check_session();
+
+	if($authorized){
+		$files = get_files();
+		render('res_delete', array('title' => "Delete Resources:",
+								  'files' => $files ));
+	} else {
+		render('res_login', array('title' => "Login", 'error' => 'you must login to view this page.'));
+	}
 });
 
 
 //post
+post('/resources/admin/login', function () {
+
+	$authorized = false;
+
+	//get login details
+
+	$username = test_input($_POST['username']);
+	$password = test_input($_POST['password']);
+
+	//create session
+
+	if (check_login($username, $password)) {
+		create_session();
+		$authorized = true;
+	}
+
+	if ($authorized){
+
+		render('res_main', array('title' => "Manage Resources"));
+
+	} else {
+	
+		render('res_login', array('title' => "Login", 'error' => 'Invalid username password combination.'));
+
+	}
+
+});
+
+
+
+
 
 post('/resources/admin/new', function () {
 
@@ -128,15 +195,54 @@ post('/resources/admin/new', function () {
 	$url = test_input($_POST['url']);
 	$file = $_FILES['file'];
 
-	//validation
+	$postdata = array('links' => array(), 'tools' => array(), 'articles' => array());
+
+	$postdata[$type]['active'] = true;
+	$postdata[$type]['title'] = $title;
+	$postdata[$type]['description'] = $description;
+	$postdata[$type]['url'] = $url;
+
+	$error = "";
 	$message = "";
 
-	 // Check file size
-	if ($file["size"] > 500000) {
-	    $message .= "Sorry, your file is too large.\n";
-	} 
+	//validation
 
-	if ($message == ""){
+	// check form type hasnt been changed
+	if (!($type == 'links' || $type == 'tools' || $type == 'articles')){
+		$error .= "Form has been tampered with.<br>";
+	}
+	// check title and desc isnt empty
+	if (empty($title) || empty($description)){
+		$error .= "Title and description are required.<br>";
+	}
+
+	if ($type == "links"){
+		// check url isnt empty and check syntax
+		if (empty($url)){
+			$error .= "Please enter a url for the link (e.g. www.example.com).<br>";
+		} elseif (!preg_match("/\b(?:(?:https?|ftp):\/\/|www\.)[-a-z0-9+&@#\/%?=~_|!:,.;]*[-a-z0-9+&@#\/%=~_|]/i",$url)){
+			$error .= "URL is not valid.<br>";	
+		} else {
+			//remove http etc from start of url
+			$url = preg_replace('#^https?://#', '', $url);
+		}
+
+	} else {
+
+		// Check file uloaded correctly
+		if ($file['size'] == 0){
+			$error .= "Please select file for upload.<br>";	
+		} elseif ($file['error'] != 0) {
+			$error .= "Error uploading file. please try again.<br>";	
+		}
+		// Check file size
+		if ($file["size"] > 5000000) {
+		    $error .= "Sorry, your file is too large. Files must be less than 5mb.<br>";
+		} 
+
+	}
+
+	if ($error == ""){
 
 		//create new file listing
 
@@ -156,13 +262,14 @@ post('/resources/admin/new', function () {
 			$extension = array_pop(explode(".", $file["name"]));
 			$path = "res/" . $type . "/files/" . $filename . "." . $extension;
 
-			if (!move_uploaded_file($file["tmp_name"], $path)) $message .= "Something went wrong uploading your file. Please try again.";
+			if (!move_uploaded_file($file["tmp_name"], $path)) $error .= "Something went wrong uploading your file. Please try again.";
 
 		}
 	}
 
 	//create file for documentation
-	if ($message == ""){
+	if ($error == ""){
+
 
 		$docspath = "res/" . $type . "/docs/" . $filename . ".txt";
 		$docsdata = $title . " #?# " . $description;
@@ -170,12 +277,13 @@ post('/resources/admin/new', function () {
 		fwrite($docsfile, $docsdata); 
 		fclose($docsfile);
 		$message = "'" . $title . "' was succesfully added to " . $type . "!";
+		$postdata = array();
 
 	} 
 
 	//render new page with message
 
-	render('res_new', array('title' => "Add resources", 'message' => $message));
+	render('res_new', array('title' => "Add resources", 'postdata' => $postdata, 'message' => $message, 'error' => $error));
 
 });
 
@@ -188,47 +296,102 @@ post('/resources/admin/edit', function () {
 	$title = test_input($_POST['title']);
 	$description = test_input($_POST['description']);
 	$url = test_input($_POST['url']);
+	$id = test_input($_POST['id']);
 	$file = $_FILES['file'];
 
 	$docspath = test_input($_POST['docspath']);
 	$filespath = test_input($_POST['filespath']);
 
-	//update files
+	$postdata = array('links' => array(), 'tools' => array(), 'articles' => array());
+
+	$postdata[$type]['active'] = true;
+	$postdata[$type]['title'] = $title;
+	$postdata[$type]['description'] = $description;
+	$postdata[$type]['url'] = $url;
+	$postdata['activeid'] = $id;
+
+	$error = "";
+	$message = "";
+
+	//validation
+
+	// check form type hasnt been changed
+	if (!($type == 'links' || $type == 'tools' || $type == 'articles')){
+		$error .= "Form has been tampered with.<br>";
+	}
+	//check file locations lead to right folder
+	if (!(substr($docspath, 0, 4) == "res/" && substr($filespath, 0, 4)) == "res/" ) {
+		$error .= "Form has been tampered with.<br>";
+	}
+
+	// check title and desc isnt empty
+	if (empty($title) || empty($description)){
+		$error .= "Title and description are required.<br>";
+	}
 
 	if ($type == "links"){
+		// check url isnt empty and check syntax
+		if (empty($url)){
+			$error .= "Please enter a url for the link (e.g. www.example.com).<br>";
+		} elseif (!preg_match("/\b(?:(?:https?|ftp):\/\/|www\.)[-a-z0-9+&@#\/%?=~_|!:,.;]*[-a-z0-9+&@#\/%=~_|]/i",$url)){
+			$error .= "URL is not valid.<br>";	
+		} else {
+			//remove http etc from start of url
+			$url = preg_replace('#^https?://#', '', $url);
+		}
 
-		$filesdata = $url;
-		$filesfile = fopen($filespath, "w");
-		fwrite($filesfile, $filesdata); 
-		fclose($filesfile);
+	} else {
 
-	} elseif ( $file['size'] != 0 ) {
-
-		$newextension = array_pop(explode(".", $file["name"]));
-
-		$noext = explode(".", $filespath);
-		$discard = array_pop($noext);
-		$noext = implode(".", $noext);
-		$newpath = $noext . "." . $newextension;
-			
-		unlink($filespath);	
-		if (!move_uploaded_file($file["tmp_name"], $newpath)) $message .= "Something went wrong uploading your file. Please try again.";
+		// Check file size
+		if ($file["size"] > 5000000) {
+		    $error .= "Sorry, your file is too large. Files must be less than 5mb.<br>";
+		} 
 
 	}
 
-	//update docs
-	if ($message == ""){
+	if ($error == ""){
+
+		//update files
+
+		if ($type == "links"){
+
+			$filesdata = $url;
+			$filesfile = fopen($filespath, "w");
+			fwrite($filesfile, $filesdata); 
+			fclose($filesfile);
+
+		} elseif ( $file['size'] != 0 ) {
+
+			$newextension = array_pop(explode(".", $file["name"]));
+
+			$noext = explode(".", $filespath);
+			$discard = array_pop($noext);
+			$noext = implode(".", $noext);
+			$newpath = $noext . "." . $newextension;
+				
+			unlink($filespath);	
+			if (!move_uploaded_file($file["tmp_name"], $newpath)) $error .= "Something went wrong uploading your file. Please try again.\n";
+
+		}
+
+	}
+
+	if ($error == ""){
+
+		//update docs
 
 		$docsdata = $title . " #?# " . $description;
 		$docsfile = fopen($docspath, "w");
 		fwrite($docsfile, $docsdata); 
 		fclose($docsfile);
 		$message = "'" . $title . "' was succesfully updated!";
+		$postdata = array();
 
 	}
+
 	//render edit page with message
 	$files = get_files();
-	render('res_edit', array('title' => "Edit resources", "files" => $files, 'message' => $message, 'test' => $test));
+	render('res_edit', array('title' => "Edit resources", "files" => $files, 'postdata' => $postdata, 'message' => $message, 'error' => $error));
 
 });
 
@@ -241,17 +404,29 @@ post('/resources/admin/delete', function () {
 	$docspath = test_input($_POST['docspath']);
 	$filespath = test_input($_POST['filespath']);
 
-	//remove files
+	$error = "";
+	$message = "";
 
-	if ( unlink($docspath) && unlink($filespath) ) 	{
-		$message = "'" . $title . "' was succesfully deleted!";
+	//validation
+
+	//check file locations lead to right folder
+	if (!(substr($docspath, 0, 4) == "res/" && substr($filespath, 0, 4)) == "res/" ) {
+		$error .= "Form has been tampered with.<br>";
 	} else {
-		$message = "Something went wrong. Please try again.";
+
+		//remove files
+
+		if ( unlink($docspath) && unlink($filespath) ) 	{
+			$message = "'" . $title . "' was succesfully deleted!";
+		} else {
+			$error = "Something went wrong. Please try again.";
+		}
+		//render edit page with message
+
 	}
-	//render edit page with message
 
 	$files = get_files();
-	render('res_delete', array('title' => "Edit resources", "files" => $files, 'message' => $message));
+	render('res_delete', array('title' => "Edit resources", "files" => $files, 'message' => $message, 'error' => $error));
 
 });
